refactor: move all blocks elements to own folders
