+++ front/src/pages/User/LessonPage/BlockElement/Delimiter/Delimiter.jsx
@@ -1,0 +1,5 @@
+import * as S from './Delimiter.styled';
+
+const Delimiter = () => <S.Delimiter />;
+
+export default Delimiter;
