+++ front/src/pages/User/LessonPage/BlockElement/Delimiter/index.js
@@ -1,0 +1,3 @@
+import Delimiter from './Delimiter';
+
+export default Delimiter;
