+++ front/src/pages/User/LessonPage/BlockElement/Embed/index.js
@@ -1,0 +1,3 @@
+import Embed from './Embed';
+
+export default Embed;
