+++ front/src/pages/User/LessonPage/BlockElement/Error/index.js
@@ -1,0 +1,3 @@
+import Error from './Error';
+
+export default Error;
