+++ front/src/pages/User/LessonPage/BlockElement/Header/Header.jsx
@@ -1,0 +1,13 @@
+import { Typography } from 'antd';
+import { BlockElementProps } from '../utils';
+
+const { Title } = Typography;
+
+const Header = ({ content }) => {
+  const { text, level } = content.data;
+  return <Title level={level}>{text}</Title>;
+};
+
+Header.propTypes = BlockElementProps;
+
+export default Header;
