+++ front/src/pages/User/LessonPage/BlockElement/Image/index.js
@@ -1,0 +1,3 @@
+import Image from './Image';
+
+export default Image;
