+++ front/src/pages/User/LessonPage/BlockElement/List/List.jsx
@@ -1,0 +1,31 @@
+import { BlockElementProps } from '../utils';
+
+const List = ({ content }) => {
+  const { style, items } = content.data;
+  if (style === 'ordered') {
+    return (
+      <ol>
+        {items.map((item, index) => (
+          // eslint-disable-next-line react/no-array-index-key
+          <li key={index}>{item}</li>
+        ))}
+      </ol>
+    );
+  }
+
+  if (style === 'unordered') {
+    return (
+      <ul>
+        {items.map((item, index) => (
+          // eslint-disable-next-line react/no-array-index-key
+          <li key={index}>{item}</li>
+        ))}
+      </ul>
+    );
+  }
+  return null;
+};
+
+List.propTypes = BlockElementProps;
+
+export default List;
