+++ front/src/pages/User/LessonPage/BlockElement/Paragraph/Paragraph.jsx
@@ -1,0 +1,14 @@
+import HtmlToReact from 'html-to-react';
+import { BlockElementProps } from '../utils';
+
+const HtmlToReactParser = HtmlToReact.Parser;
+
+const Paragraph = ({ content }) => {
+  const htmlInput = content?.data?.text;
+  const htmlToReactParser = new HtmlToReactParser();
+  return htmlToReactParser.parse(htmlInput);
+};
+
+Paragraph.propTypes = BlockElementProps;
+
+export default Paragraph;
