+++ front/src/pages/User/LessonPage/BlockElement/Paragraph/index.js
@@ -1,0 +1,3 @@
+import Paragraph from './Paragraph';
+
+export default Paragraph;
