+++ front/src/pages/User/LessonPage/BlockElement/Quote/index.js
@@ -1,0 +1,3 @@
+import Quote from './Quote';
+
+export default Quote;
