+++ front/src/pages/User/LessonPage/BlockElement/Table/Table.styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const CustomTableWrapper = styled.div`
+  margin-top: 1rem;
+  background-color: white;
+  padding: 1rem;
+  border-radius: 1rem;
+`;
+
+export const CustomTable = styled.table`
+  table-layout: fixed;
+  width: 100%;
+  border-collapse: collapse;
+  td {
+    padding: 19px;
+  }
+`;
