+++ front/src/pages/User/LessonPage/BlockElement/Table/index.js
@@ -1,0 +1,3 @@
+import Table from './Table';
+
+export default Table;
